// src/app/helpers/weapon-default-skin-infos.enum.ts
export enum weaponDefaultSkinUUIDs {
  classic = "24aee897-4cdc-b0fd-e596-1ba90fa6d1b2",
  shorty = "48ad078a-4dae-2b85-a945-f4b6d1efecbb",
  frenzy = "f06657f3-48b6-6314-7235-a9a2749df5b9",
  ghost = "1c63b43b-43c4-04e4-01c9-7aa1bffa5ac1",
  sheriff = "1ef6ba68-4dbe-30c7-6bc8-93a6c6f13f04",
  stinger = "940fb417-4a9c-3004-41f5-3e8f1f4178b2",
  spectre = "f01d1307-4299-42f5-2c5e-7dab7e69ab19",
  bucky = "70c97fb2-4d79-d4bb-5173-a1888cd4bfd9",
  judge = "acd26127-48ff-8b9e-7ba6-b989af8a4b24",
  bulldog = "724a7f42-4315-eccf-0e76-77bdd3ec2e09",
  guardian = "3bf1e8e0-47e8-f27a-6054-929575f41a54",
  phantom = "337cb216-4a6e-d85d-88c2-f29ab317784c",
  vandal = "27f21d97-4c4b-bd1c-1f08-31830ab0be84",
  marshal = "fd44b2d5-49ee-77ab-fa56-588f3ac0c268",
  operator = "d1f2920f-469a-3431-ad96-96afbd0017f2",
  ares = "5305d9c4-4f46-fbf4-9e9a-dea772c263b5",
  odin = "f454efd1-49cb-372f-7096-d394df615308",
}

export enum weaponDefaultSkinImages {
  classic = "https://media.valorant-api.com/weaponskinchromas/4b2d5b4f-4955-4208-286c-abadec250cdd/fullrender.png",
  shorty = "https://media.valorant-api.com/weaponskinchromas/95608504-4c8b-1408-1612-0f8200421c49/fullrender.png",
  frenzy = "https://media.valorant-api.com/weaponskinchromas/dc99ed5a-4d75-87a0-c921-75963ea3c1e1/fullrender.png",
  ghost = "https://media.valorant-api.com/weaponskinchromas/947a28b6-4e0f-61fb-e795-bc9a5e7b7129/fullrender.png",
  sheriff = "https://media.valorant-api.com/weaponskinchromas/5a59bd61-48a9-af61-c00f-4aa21deca9a8/fullrender.png",
  stinger = "https://media.valorant-api.com/weaponskinchromas/31bb2115-4c62-d37c-43c4-11b8fee7f212/fullrender.png",
  spectre = "https://media.valorant-api.com/weaponskinchromas/a9aaccca-4cdc-02ea-1d7e-89bbacecc0e2/fullrender.png",
  bucky = "https://media.valorant-api.com/weaponskinchromas/3d8ffcfe-4786-0180-42d7-e1be18dd1cab/fullrender.png",
  judge = "https://media.valorant-api.com/weaponskinchromas/b71ae8d6-44bb-aa4c-0d2a-dc9ed9e66410/fullrender.png",
  bulldog = "https://media.valorant-api.com/weaponskinchromas/bf35f404-4a14-6953-ced2-5bafd21639a0/fullrender.png",
  guardian = "https://media.valorant-api.com/weaponskinchromas/0f934388-418a-a9e7-42a7-21b27402e46c/fullrender.png",
  phantom = "https://media.valorant-api.com/weaponskinchromas/52221ba2-4e4c-ec76-8c81-3483506d5242/fullrender.png",
  vandal = "https://media.valorant-api.com/weaponskinchromas/19629ae1-4996-ae98-7742-24a240d41f99/fullrender.png",
  marshal = "https://media.valorant-api.com/weaponskinchromas/1afec971-4170-f29b-1c94-07a0eff270ab/fullrender.png",
  operator = "https://media.valorant-api.com/weaponskinchromas/4914f50d-49f9-6424-ca80-9486c45a138d/fullrender.png",
  ares = "https://media.valorant-api.com/weaponskinchromas/b33de820-4061-8b85-31ce-808f1a2c58f5/fullrender.png",
  odin = "https://media.valorant-api.com/weaponskinchromas/2f93861d-4b2f-2175-af0c-3ba0c736e257/fullrender.png",
}