// src/app/helpers/weapon-names.enum.ts
export enum weaponNames {
  classic = "Classic",
  shorty = "Shorty",
  frenzy = "Frenzy",
  ghost = "Ghost",
  sheriff = "Sheriff",
  stinger = "Stinger",
  spectre = "Spectre",
  bucky = "Bucky",
  judge = "Judge",
  bulldog = "Bulldog",
  guardian = "Guardian",
  phantom = "Phantom",
  vandal = "Vandal",
  marshal = "Marshal",
  operator = "Operator",
  ares = "Ares",
  odin = "Odin",
}